% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expomedt.R
\name{FuncExit}
\alias{FuncExit}
\title{End the module analysis}
\usage{
FuncExit(PID)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by InitMedt.}
}
\description{
End the module analysis
}
\examples{
res <- InitMedt()
   res1 <- LoadMedt(PID = res$PID, UseExample = "example#1", DataPath = NULL,
   VocaPath = NULL)
   res2 <- XMlists(PID = res$PID)
   FuncExit(PID = res$PID)
}
\author{
Bin Wang (corresponding author)
}
