% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expomedt.R
\name{ImptM}
\alias{ImptM}
\title{Estimate the importance of mediators}
\usage{
ImptM(PID, OutPath, VarsY, ...)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated
by InitMedt.}

\item{OutPath}{chr. Output file directory, e.g. "D:/output". If "default",
the current working directory will be set.It should be noted that the
slash symbol is "/", not "\".}

\item{VarsY}{chr. The outcome variable. Either continuous, binary or count
type is permitted.}

\item{VarsX}{chr. Exposure variables included in estimation procedure.
When "default" is specified, all exposure variables in the data
will be used. The shirnkaged exposure variables are also permitted.
It should be noted that there is fixed format for the characters separated
with comma and without space, e.g., "X1,X2,X3".}

\item{VarsC}{chr. Covariates included in estimation procedure.
When "default" is specified, all covariates in the data will
be used. It should be noted that there is fixed format for the characters
separated with comma and without space, e.g., "C1,C2".}
}
\value{
A list containing four elements where the raw estimation results
as well as the tidy tables for display are stored. The elements of
that list include:
\enumerate{
\item "MedtImptM_all": the raw estimation result where the importance of
each mediator was evaluated among whole mediators.
\item "MedtImptM_list": the raw estimation result where the importance of
each mediator was evaluated among each mediator group.
\item "MedtImptM_table1": the tidy table for MedtImptM_all and MedtImptM_list,
where significant estimations (q value <0.2) are expressed with an asterisk
mark (*).
\item "MedtImptM_table2": the tidy table for MedtImptM_all and MedtImptM_list,
where only significant estimations (q value <0.2) are displayed in this table.
MedtImptM_table2 is as same content as MedtImptM_table1.
}
}
\description{
Estimate the importance of mediators.
}
\details{
ImptM function estimates the importance of mediators among given exposures.
The estimation procedure is mainly realized by the hima function in HIMA package.
ImptM also provides another evaluation method by the bama function in bama package.
The users can search for these packages for further information.
}
\examples{
res <- InitMedt()
   res1 <- LoadMedt(PID = res$PID, UseExample = "example#1", DataPath = NULL,
   VocaPath = NULL)
   res2 <- XMlists(PID = res$PID)
   res3 <- ImptM(PID = res$PID, VarsY = "Y1",
   VarsX = "default", VarsC = "default")
   res4 <- ImptM(PID = res$PID, VarsY = "Y1",
   VarsX = "X1,X2,X3", VarsC = "C1")
}
\author{
Mengyuan Ren, Bin Wang(corresponding author)
}
