% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expomedt.R
\name{LoadMedt}
\alias{LoadMedt}
\title{Load data file for Mediation module}
\usage{
LoadMedt(PID = PID, UseExample = "default", ...)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by
InitMedt.}

\item{UseExample}{chr. Method of uploading data. If "default"，
user should upload their own data files, or use "example#1"
provided by this module.}

\item{DataPath}{chr. Input data file directory, e.g.
"D:/test/eg_Medt_data.xlsx". It should be noted that the slash
symbol is "/", not "\".}

\item{VocaPath}{chr. Input vocabulary file directory, e.g.
"D:/test/eg_Medt_data.xlsx". It should be noted that the slash
symbol is "/", not "\".}
}
\value{
An R6 class object containing the input data and vocabulary file.
}
\description{
Load data file for Mediation module
}
\details{
LoadMedt function loads the data file and the vocabulary file into
the R6 object that InitMedt created. Noted that there are several data
format requirments for the data and vocabulary file.
For data file, the first three column must be named as "SampleID", "SubjectID"
and "Group" in sequence. The "Gourp" variable should be a character variable
to category data into two groups: "train" and "test" group. Outcome variables
should be named as "Y*". e.g., Y1, Y2, Y3... Similarly, exposure variables
should be named as "X*". e.g., X1, X2, X3..., and mediator variables should
be named as "M*". e.g., M1, M2, M3...
For vocabulary file, the first column should be a character variable named
"SerialNo" indicating the names of outcome, exposure and mediator variables
in the data file. These names should be consistent with the variable names
in data file. The second column should be a character variable named "FullName"
indicating the full names (labels) of the variables. The third column should
be a character variable named "SubgroupName" indicating the groups the exposure
or mediator variables belong to.
}
\examples{
res <- InitMedt()
   res1 <- LoadMedt(PID = res$PID, UseExample = "example#1", DataPath = NULL,
   VocaPath = NULL)
}
\author{
Mengyuan Ren, Bin Wang(corresponding author)
}
