% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expomedt.R
\name{Pairwise}
\alias{Pairwise}
\title{Implement pairwise mediation analyses}
\usage{
Pairwise(PID, OutPath, VarsY, Family, ...)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated
by InitMedt.}

\item{OutPath}{chr. Output file directory, e.g. "D:/output". If "default", the
current working directory will be set.It should be noted that the slash
symbol is "/", not "\".}

\item{VarsY}{chr. The outcome variable. Either continuous, binary or count
type is permitted.}

\item{VarsX}{chr. Exposure variables included in pairwise mediation modelling.
When "default" is specified, all exposure variables in the data will
be used. It should be noted that there is fixed format for the characters
separated with comma and without space, e.g., "X1,X2,X3".}

\item{VarsM}{chr. Mediator variables included in pairwise mediation modelling.
When "default" is specified, all mediator variables in the data will
be used. It should be noted that there is fixed format for the characters
separated with comma and without space, e.g., "M1,M2,M3".}

\item{VarsC}{chr. Covariates included in pairwise mediation modelling.
When "default" is specified, all covariates in the data will
be used. It should be noted that there is fixed format for the characters
separated with comma and without space, e.g., "C1,C2".}

\item{Family}{chr. The error distribution and link function to be used in the
model. Available options include "linear" and "gaussian" for continuous
outcome variables, and "binomial" as well as "poisson" options for binary
and count outcome variables.}

\item{Iter}{num. The number of iteration times. Default is 500. To get a more stable
result, a minimum iteration of 1,000 is recommended. To obtain more digits
of P value of parameter estimation, a minimum iteration of 5,000 is suggested.}
}
\value{
A list containing one dataframe that contains the pairwise mediation modelling results.
\enumerate{
\item "MedtPairWise_Stats": pairwise mediation modelling results.
}
}
\description{
Implement pairwise mediation analyses for each pair of exposure
and mediator.
}
\details{
Pairwise function implements mediation modelling for each pair of exposure
and mediator. Given M exposures and N mediators, an exhaustive rule will be executed
and M*N pair-wised mediation modelling are fitted. The modelling was realized using
mediate function in mediation package.
}
\examples{
res <- InitMedt()
   res1 <- LoadMedt(PID = res$PID, UseExample = "example#1", DataPath = NULL,
   VocaPath = NULL)
   res2 <- XMlists(PID = res$PID)
   res3 <- Pairwise(PID=res$PID, VarsY = "Y1",
   VarsX = "default", VarsM = "default", VarsC = "default", Family = "linear",
   Iter = 500)
   res4 <- Pairwise(PID=res$PID, VarsY = "Y1",
   VarsX = "X1,X2,X3", VarsM = "M1,M2,M3", VarsC = "C1", Family = "linear",
   Iter = 500)
}
\author{
Mengyuan Ren, Bin Wang(corresponding author)
}
