% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expomedt.R
\name{RedM}
\alias{RedM}
\title{Mediator dimension reduction}
\usage{
RedM(PID, OutPath, VarsY, Family, ...)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated
by InitMedt.}

\item{OutPath}{chr. Output file directory, e.g. "D:/output". If "default",
the current working directory will be set.It should be noted that the
slash symbol is "/", not "\".}

\item{VarsY}{chr. The outcome variable. Either continuous, binary or count
type is permitted.}

\item{VarsX}{chr. Exposure variables included in mediation modelling.
When "default"(recommended) is specified, all exposure variables in the data
will be used. It should be noted that there is fixed format for the characters
separated with comma and without space, e.g., "X1,X2,X3".}

\item{VarsC}{chr. Covariates included in mediation modelling.
When "default" is specified, all covariates in the data will
be used. It should be noted that there is fixed format for the characters
separated with comma and without space, e.g., "C1,C2".}

\item{Method}{chr. Dimension reduction method. Available options include
"mean" and "pdm1"(default). "mean" option is recommended when outcome variable
is a binary variable because in some situations error might occur in "pdm1"
settings for binary outcome.}

\item{Family}{chr. The error distribution and link function to be used in the
model. Available options include "linear" and "gaussian" for continuous
outcome variables, and "binomial" as well as "poisson" options for binary
and count outcome variables.}

\item{Iter}{num. The number of iteration times. Default is 500. To get a more stable
result, a minimum iteration of 1,000 is recommended. To obtain more digits
of P value of parameter estimation, a minimum iteration of 5,000 is suggested.}
}
\value{
A list containing two elements where the mediator dimension reduction
modelling results with exposures are stored. That list include:
\enumerate{
\item "MedtRedM_all": a dataframe containing the mediator dimension reduction
result where all mediators are considered as one group and that shrinkaged
mediator is paired and modelled with each exposure.
\item "MedtRedM_list": a dataframe containing the mediator dimension reduction
result where mediators are shrinkaged in their own subgroups, and these
shrinkaged mediators are paired and modelled with each exposure.
}
}
\description{
Implement mediator dimension reduction for mediators.
}
\details{
RedM function provides two alternative methods for mediator dimension
reduction, including sum method as well as pdm1 method via
PDM package. By default, all mediators will be included for dimension reduction.
Afterwards, mediation models will be built between given exposures as well as
shrinkaged mediator variables.
}
\examples{
res <- InitMedt()
   res1 <- LoadMedt(PID = res$PID, UseExample = "example#1", DataPath = NULL,
   VocaPath = NULL)
   res2 <- XMlists(PID = res$PID)
   res3 <- RedM(PID=res$PID, VarsY = "Y1", VarsX = "default",
   VarsC = "default", Method = "mean", Family = "linear", Iter = 500)
   res4 <- RedM(PID=res$PID, VarsY = "Y1", VarsX = "X1,X2,X3",
   VarsC = "C1", Method = "mean", Family = "linear", Iter = 500)
}
\author{
Mengyuan Ren, Bin Wang(corresponding author)
}
