% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expomedt.R
\name{RedX}
\alias{RedX}
\title{Expoures dimension reduction}
\usage{
RedX(PID, OutPath, VarsY, Family, ...)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated
by InitMedt.}

\item{OutPath}{chr. Output file directory, e.g. "D:/output". If "default",
the current working directory will be set.It should be noted that the
slash symbol is "/", not "\".}

\item{VarsY}{chr. The outcome variable. Either continuous, binary or count
type is permitted.}

\item{VarsC}{chr. Covariates included in mediation modelling.
When "default" is specified, all covariates in the data will
be used. It should be noted that there is fixed format for the characters
separated with comma and without space, e.g., "C1,C2".}

\item{VarsM}{chr. Mediator variables included in pairwise mediation modelling.
When "default" is specified, all mediator variables in the data will
be used. It should be noted that there is fixed format for the characters
separated with comma and without space, e.g., "M1,M2,M3".}

\item{Method}{chr. Dimension reduction method. Available options include
"gcdnet" and "mean"(default). "mean" option is recommended when outcome variable
is a binary variable because in some situations low variance of shrinkaged exposure
variable might obtain in "gcdnet" settings for binary outcome.}

\item{Folds}{num. Number of cross validation for gcdnet method. Default is 10.}

\item{Family}{chr. The error distribution and link function to be used in the
model. Available options include "linear" and "gaussian" for continuous
outcome variables, and "binomial" as well as "poisson" options for binary
and count outcome variables.}

\item{Iter}{num. The number of iteration times. Default is 500. To get a more stable
result, a minimum iteration of 1,000 is recommended. To obtain more digits
of P value of parameter estimation, a minimum iteration of 5,000 is suggested.}
}
\value{
A list containing three elements where the exposure dimension reduction
variables as well as their mediation modelling results with mediators
are stored. That list include:
\enumerate{
\item "MedtRedX_ERSall": a dataframe containing the exposure dimension reduction
result where all exposures are considered as one group (ERS_All).
\item "MedtRedX_ERSlist": a dataframe containing the exposure dimension reduction
results where exposures are shrinkaged in their own subgroups.
\item "MedtRedX_Stats": a dataframe containing the mediation modelling results
where the shrinkaged exposure variables were paired with each mediator provided.
}
}
\description{
Implement dimension reduction for exposures.
}
\details{
RedX function provides two alternative methods for exposure dimension
reduction, including sum method as well as adaptive elastic net method via
gcdnet package. By default, all exposures will be included for dimension reduction.
Afterwards, mediation models will be built between given mediators as well as
shrinkaged exposure variables.
}
\examples{
res <- InitMedt()
   res1 <- LoadMedt(PID = res$PID, UseExample = "example#1", DataPath = NULL,
   VocaPath = NULL)
   res2 <- XMlists(PID = res$PID)
   res3 <- RedX(PID=res$PID, VarsY = "Y1",
   VarsC = "default", VarsM = "default", Method = "mean", Folds = 10,
   Family = "linear", Iter = 500)
   res4 <- RedX(PID=res$PID, VarsY = "Y1",
   VarsC = "C1", VarsM = "M1,M2,M3", Method = "mean", Folds = 10,
   Family = "linear",Iter = 500)
}
\author{
Mengyuan Ren, Bin Wang(corresponding author)
}
