% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expomedt.R
\name{VizRedXM}
\alias{VizRedXM}
\title{Visualize RedXM mediation modelling result}
\usage{
VizRedXM(PID, OutPath, ...)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated
by InitMedt.}

\item{OutPath}{chr. Output file directory, e.g. "D:/output". If "default",
the current working directory will be set.It should be noted that the
slash symbol is "/", not "\".}

\item{Brightness}{chr. Visualization brightness. Available options include
"Light" and "Dark".}

\item{Pallette}{chr. Visualization palette. Available options include "default1",
"default2" and "Journal". The "Journal" option provides several journal
preference styles including cell, nature, science, lancet, nejm, and jama.}
}
\value{
A list containing two elements where the visualized exposure and
mediator dimension reduction plot as well as the organized plotting data
are stored. The elements of that list include:
\enumerate{
\item "MedtRedXM_plotadta": a dataframe containing the organized plotting
data extracted from exposure and mediator dimension reduction result.
\item "MedtRedXM_plot": a visualized plot for exposure and mediator
dimension reduction result.
}
}
\description{
Visualize the RedXM mediation result. It should be noted that
the RedXM mediation modelling result has been built by RedXM function
prior to using it.
}
\details{
VizRedXM draws a visualized display for RedXM results.
Piror to using VizMedtPair, make sure that the users have built mediation
models by RedXM function.
}
\examples{
res <- InitMedt()
   res1 <- LoadMedt(PID = res$PID, UseExample = "example#1", DataPath = NULL,
   VocaPath = NULL)
   res2 <- XMlists(PID = res$PID)
   res3 <- RedX(PID=res$PID, VarsY = "Y1",
   VarsC = "default", VarsM = "default", Method = "mean", Folds = 10,
   Family = "linear", Iter = 500)
   res4 <- RedXM(PID=res$PID, VarsY = "Y1",
   VarsC = "default", Method = "mean", Family = "linear", Iter = 500)
   res5 <- VizRedXM(PID=res$PID, Brightness = "bright",
   Pallette = "nature")
}
\author{
Mengyuan Ren, Bin Wang(corresponding author)
}
