% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expomedt.R
\name{XMlists}
\alias{XMlists}
\title{Divide exposures and mediators into different groups}
\usage{
XMlists(PID, OutPath)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated
by InitMedt.}

\item{OutPath}{chr. Output file directory, e.g. "D:/output". If "default",
the current working directory will be set.It should be noted that the
slash symbol is "/", not "\".}
}
\value{
A list containing two lists where exposure and mediator variables
were respectively categorized into subgroups in the form of dataframe.
The elements of that list include:
\enumerate{
\item "ExpoList": a list containing several dataframes that represent various exposure subgroups.
\item "MediList": a list containing several dataframes that represent various mediator subgroups.
}
}
\description{
Divide exposures and mediators into different groups according
to the information specified in vocabulary file.
}
\details{
XMlists divides all exposures and mediators into different groups.
Therefore, the group information(i.e., the "Subgroup" variable) in vocabulary
file is essential for XMlists function. Before using XMlist, the users must
provide the information in advance and upload it by LoadMedt function.
}
\examples{
res <- InitMedt()
   res1 <- LoadMedt(PID = res$PID, UseExample = "example#1", DataPath = NULL,
   VocaPath = NULL)
   res2 <- XMlists(PID = res$PID)
}
\author{
Mengyuan Ren, Bin Wang(corresponding author)
}
